==> 01_ClinVar_VCF_parser.ipynb <==
# **ClinVar VCF Parser**
### Author: Daniel Brock
### Date: 7/1/2023
### Purpose: to select pathogenic and likely pathogenic (PLP) and benign and likely benign (BLB) SNVs from ClinVar after 2020-11-21 (date of MutScore paper cutoff)

# Importing required packages
import io
import os
import pandas as pd

# Working directory
cwd = os.getcwd()
print(cwd)

def read_vcf(path):
    """
    Function to parse a vcf file as a dataframe.
    Source: https://gist.github.com/dceoy/99d976a2c01e7f0ba1c813778f9db744
    Input: a file path to a VCF file from ClinVar
    Output: a pandas dataframe
    """
    with open(path, 'r') as f:
        lines = [l for l in f if not l.startswith('##')]
    return pd.read_csv(
        io.StringIO(''.join(lines)),
        dtype={'#CHROM': str, 'POS': int, 'ID': str, 'REF': str, 'ALT': str, 'QUAL': str, 'FILTER': str, 'INFO': str},
        sep='\t'
    ).rename(columns={'#CHROM': 'CHROM'})

# Importing the ClinVar VCF file (as of 2023-06-26)
# Source: https://ftp.ncbi.nlm.nih.gov/pub/clinvar/
clinvar = read_vcf(path=cwd+"/ClinVar_Data/clinvar_20230626.vcf")
display(clinvar.shape)
display(clinvar.head())

clinvar["INFO"][0]

## Adding new columns based on ClinVar's "INFO" column

# ClinVar Allele ID
pattern = r"ALLELEID=(.*?);"
clinvar["ALLELEID"] = clinvar["INFO"].str.extract(pattern)
print("Done with ALLELEID")

# tag-value pairs of disease database name and identifier
pattern = r"CLNDISDB=(.*?);"
clinvar["CLNDISDB"] = clinvar["INFO"].str.extract(pattern)
print("Done with CLNDISDB")

# ClinVar's preferred disease name for the concept specified by disease identifiers in CLNDISDB
pattern = r"CLNDN=(.*?);"
clinvar["CLNDN"] = clinvar["INFO"].str.extract(pattern)
print("Done with CLNDN")

# top-level (primary assembly, alt, or patch) HGVS expression
pattern = r"CLNHGVS=(.*?);"
clinvar["CLNHGVS"] = clinvar["INFO"].str.extract(pattern)
print("Done with CLNHGVS")

# ClinVar's review status for the Variation ID
pattern = r"CLNREVSTAT=(.*?);"
clinvar["CLNREVSTAT"] = clinvar["INFO"].str.extract(pattern)
print("Done with CLNREVSTAT")

# clinical significance for this single variant; multiple values are separated by a vertical bar
pattern = r"CLNSIG=(.*?);"
clinvar["CLNSIG"] = clinvar["INFO"].str.extract(pattern)
print("Done with CLNSIG")

# variant type
pattern = r"CLNVC=(.*?);"
clinvar["CLNVC"] = clinvar["INFO"].str.extract(pattern)
print("Done with CLNVC")

# Sequence Ontology ID for the variant type (www.sequenceontology.org)
pattern = r"CLNVCSO=(.*?);"
clinvar["CLNVCSO"] = clinvar["INFO"].str.extract(pattern)
print("Done with CLNVCSO")

# gene(s) for the variant reported as gene symbol:NCBI GeneID. The gene symbol and ID are delimited by a colon and each pair is delimited by a vertical bar.
pattern = r"GENEINFO=(.*?);"
clinvar["GENEINFO"] = clinvar["INFO"].str.extract(pattern)
clinvar[["Gene", "Gene_ID"]] = clinvar["GENEINFO"].str.split(":", 1, expand=True)
print("Done with GENEINFO")

# comma separated list of molecular consequence in the form of Sequence Ontology ID|molecular_consequence
pattern = r"MC=(.*?);"
clinvar["MC"] = clinvar["INFO"].str.extract(pattern)
print("Done with MC")

# Allele origin reported to ClinVar
# 0 = unknown
# 1 = germline
# 2 = somatic 
# 4 = inherited
# 8 = paternal 
# 16 = maternal 
# 32 = de-novo
# 64 = biparental
# 128 = uniparental 
# 256 = not-tested
# 512 = tested-inconclusive
# 1073741824 = other
pattern = r"ORIGIN=(.*?);"
clinvar["ORIGIN"] = clinvar["INFO"].str.extract(pattern)
print("Done with ORIGIN")

display(clinvar.head())

## Filtering for BLB and PLP for any SNV reported to ClinVar after 2020-11-21 (MutScore training cutoff)

clinvar.columns

clinvar["CLNVC"].unique()

clinvar["CLNSIG"].unique()

# Filtering for SNVs
clinvar_filt = clinvar[clinvar["CLNVC"]=="single_nucleotide_variant"]
print("Clinvar unfiltered:", clinvar.shape)
print("Filtered for SNVs:", clinvar_filt.shape)

# Filtering out dates AFTER 2020-11-21
clinvar_filt["ID"] = clinvar_filt["ID"].astype(int)
clinvar_filt = clinvar_filt.sort_values(by="ID", ascending=False)
clinvar_filt = clinvar_filt[clinvar_filt["ID"] > 986732]  #if you look up ID 986732 in clinvar, this is the first SNV reported to clinvar after 2020-11-21
print("Filtered for dates after 2020-11-21:", clinvar_filt.shape)

# Filtering for variants with PLP and BLB annotations only 
pathogenecity = ["Benign", "Likely_benign", "Benign/Likely_benign", "Pathogenic", "Likely_pathogenic", "Pathogenic/Likely_pathogenic"]
clinvar_filt = clinvar_filt[clinvar_filt["CLNSIG"].isin(pathogenecity)]
print("Filtered for pathogenicity annotations:", clinvar_filt.shape)

# Cleaning and mapping data 

# Mapping and binarizing pathogenicity
clinvar_filt["PATHOGENICITY"] = clinvar_filt["CLNSIG"].map(
    {"Benign":"BLB", "Likely_benign":"BLB", "Benign/Likely_benign":"BLB", "Pathogenic":"PLP", "Likely_pathogenic":"PLP", "Pathogenic/Likely_pathogenic":"PLP"})
clinvar_filt["y_test"] = clinvar_filt["PATHOGENICITY"].map({"PLP": 1, "BLB": 0})

# Renaming chromosomes for ANNOVAR input
chromosomes = {'1':'chr1', '2':'chr2', '3':'chr3', '4':'chr4', '5':'chr5', 
                '6':'chr6', '7':'chr7', '8':'chr8', '9':'chr9', '10':'chr10', 
                '11':'chr11', '12':'chr12', '13':'chr13', '14':'chr14', '15':'chr15', 
                '16':'chr16', '17':'chr17', '18':'chr18', '19':'chr19', '20':'chr20', 
                '21':'chr21', '22':'chr22', 'X':'chrX', 'Y':'chrY', 'MT':'M'}
clinvar_filt["CHROM"] = clinvar_filt["CHROM"].map(chromosomes)

# Making the end position the same name as the start position for SNVs and reordering/renaming columns for ANNOVAR input
clinvar_filt["END"] = clinvar_filt["POS"]
clinvar_filt = clinvar_filt[['CHROM', 'POS', 'END', 'REF', 'ALT', 'QUAL', 'FILTER', 'INFO', 'ID', 'ALLELEID', 'CLNDISDB', 'CLNDN', 'CLNHGVS', 'CLNREVSTAT', 'CLNVC', 'CLNVCSO', 'GENEINFO', 'Gene', 'Gene_ID', 'MC', 'ORIGIN', 'CLNSIG', 'PATHOGENICITY', 'y_test']]
clinvar_filt.columns = ['CHROM', 'START', 'END', 'REF', 'ALT', 'QUAL', 'FILTER', 'INFO', 'ID', 'ALLELEID', 'CLNDISDB', 'CLNDN', 'CLNHGVS', 'CLNREVSTAT', 'CLNVC', 'CLNVCSO', 'GENEINFO', 'Gene', 'Gene_ID', 'MC', 'ORIGIN', 'CLNSIG', 'PATHOGENICITY', 'y_test']

# Displaying the filtered df
display(clinvar_filt.shape)
display(clinvar_filt.head())

clinvar_filt.columns

clinvar_filt["CLNSIG"].value_counts()

clinvar_filt["PATHOGENICITY"].value_counts()

clinvar_filt.dtypes

## Exporting as a txt file with the extention "avinput"

#clinvar_filt.to_csv(cwd+'/annovar_files/clinvar_filt_20201128-20230626.avinput', sep="\t", index=False, header=False)

### Optional filtering using only RetNet genes

# Importing retnet genes
retnet = pd.read_excel(cwd+"/RetNet/RetNet_AD-AR_FINALIZED.xlsx", sheet_name="gene_info")
retnet_genes = retnet["gene"]

# Filtering filtered ClinVar variants to only include RetNet Genes
clinvar_filt_retnet = clinvar_filt[clinvar_filt["Gene"].isin(retnet_genes)]
display(clinvar_filt_retnet.shape)
display(clinvar_filt_retnet.head())

len(clinvar_filt_retnet["Gene"].unique())

## Exporting filtered RetNet SNVs as a txt file with the extention "avinput"

clinvar_filt_retnet.to_csv(cwd+'/annovar_files/clinvar_filt_retnet.avinput', sep="\t", index=False, header=False)
