==> 05_classifier_thresholds.ipynb <==
# **Classifier Thresholds for Likely Benign and Likely Pathogenic SNVs** 

### Author: Daniel Brock
### Date: 7/15/2023
### Purpose: to define the upper and lower threshold value for each model (percent at while predictor is 95% accurate)

# Importing required packages
import os
import pandas as pd
import numpy as np

# Setting working directory
cwd = os.getcwd()
print(cwd)

## **Defining Functions used to calculate upper and lower threshold values for classifiers**

def likely_pathogenic(min_score, max_score, df, classifier, iterations=1001):
    """
    Function to calculate the upper threshold of a classifier model, where 95% of variants from the training ClinVar set are correctly classified as PLP.
    Inputs:
    - min_score: minimum score of a classifier
    - max_score: maximum score of a classifier
    - df: dataframe containing variants, classifier scores, and pathogenicity classifications
    - classifier: name of the classifier model
    Outputs:
    - scores dataframe: percent_PLP and percent_BLB for each scoring option
    - LP_threshold: threshold for a variant being likely pathogenic (LP)
    """
    
    # Creating an array of scores in the thousandths decimal place (used by classifiers)
    scores = np.linspace(min_score, max_score, num=iterations)
    scores = np.around(scores, 3)
    
    # Looping through the ClinVar variants data to get percent PLP and percent BLB for each possible classifier score
    df2 = df[df[classifier].notna()]
    scores_df_list = []
    for score in scores:
        s_df = df2[df2[classifier] >= score]
        plps = s_df[s_df["PATHOGENICITY"]=="PLP"].shape[0] #number of PLP variants classified by clinvar
        blbs = s_df[s_df['PATHOGENICITY']=="BLB"].shape[0] #number of BLB variants classified by clinvar
        pb = pd.DataFrame({'PLP': [plps], 'BLB': [blbs]})
        pb["percent_PLP"] = pb["PLP"] / (pb["PLP"] + pb["BLB"]) * 100
        pb["percent_BLB"] = pb["BLB"] / (pb["PLP"] + pb["BLB"]) * 100
        pb[classifier] = score
        scores_df_list.append(pb)
    scores_df = pd.concat(scores_df_list, axis=0)
    
    # Returning scores df and likely pathogenic threshold
    try:
        lp = min(scores_df[scores_df["percent_PLP"] > 95][classifier])
    except:
        lp = np.nan
        print("Had Issues")
    return scores_df, lp

def likely_benign(min_score, max_score, df, classifier, iterations=1001):
    """
    Function to calculate the lower threshold of a classifier model, where 95% of variants from the training ClinVar set are correctly classified as BLB.
    Inputs:
    - min_score: minimum score of a classifier
    - max_score: maximum score of a classifier
    - df: dataframe containing variants, classifier scores, and pathogenicity classifications
    - classifier: name of the classifier model
    Outputs:
    - scores dataframe: percent_PLP and percent_BLB for each scoring option
    - LP_threshold: threshold for a variant being likely benign (LB)
    """
    
    # Creating an array of scores in the thousandths decimal place (used by classifiers)
    scores = np.linspace(min_score, max_score, num=iterations)
    scores = np.around(scores, 3)
    
    # Looping through the ClinVar variants data to get percent PLP and percent BLB for each possible classifier score
    df2 = df[df[classifier].notna()]
    scores_df_list = []
    for score in scores:
        s_df = df2[df2[classifier] <= score]
        plps = s_df[s_df["PATHOGENICITY"]=="PLP"].shape[0] #number of PLP variants classified by clinvar
        blbs = s_df[s_df['PATHOGENICITY']=="BLB"].shape[0] #number of BLB variants classified by clinvar
        pb = pd.DataFrame({'PLP': [plps], 'BLB': [blbs]})
        pb["percent_PLP"] = pb["PLP"] / (pb["PLP"] + pb["BLB"]) * 100
        pb["percent_BLB"] = pb["BLB"] / (pb["PLP"] + pb["BLB"]) * 100
        pb[classifier] = score
        scores_df_list.append(pb)
    scores_df = pd.concat(scores_df_list, axis=0)
    
    # Returning scores df and likely pathogenic threshold
    lp = max(scores_df[scores_df["percent_BLB"] > 95][classifier])
    return scores_df, lp

# **Getting Threshold Values for ALL untrained ClinVar RetNet Genes**

### Source: Clinvar VCF File: https://www.ncbi.nlm.nih.gov/clinvar/

rc = pd.read_excel(cwd+"/annovar_files/clinvar_filt_retnet_annotated_PLP-BLB_Nonsynonomous_Variants.xlsx", na_values=".")
display(rc.shape)
display(rc.head())

# Skip if done with preprocessed data

# Importing annotated clinvar retnet variants
#rc = pd.read_csv(cwd+"/annotated_clinvar_filt_retnet_20230626.hg19_multianno.csv", na_values=".")

# Importing the clinvar retnet-filtered avinput file, which as the pathogenicity scores
#clin_avi = pd.read_table(cwd+"/clinvar_filt_retnet_20230626.avinput", sep="\t",
#                        names=['Chr', 'Start', 'End', 'Ref', 'Alt', 'QUAL', 'FILTER', 'INFO', 'ID', 'ALLELEID', 'CLNDISDB', 'CLNDN', 'CLNHGVS', 'CLNREVSTAT', 'CLNVC', 'CLNVCSO', 'GENEINFO', 'Gene', 'Gene_ID', 'MC', 'ORIGIN', 'CLNSIG', 'Pathogenicity', 'y_test'])
#clin_avi = clin_avi[['Chr', 'Start', 'End', 'Ref', 'Alt', 'Pathogenicity', 'y_test']]
#display(clin_avi.shape)
#display(clin_avi.head())

# Merging to get pathogenicity rankings
#rc = pd.merge(rc, clin_avi, on=['Chr', 'Start', 'End', 'Ref', 'Alt'])

# Filtering for nonsynonymous SNVs
#rc = rc[rc["ExonicFunc.refGene"]=="nonsynonymous SNV"]
#display(rc.shape)
#display(rc.head())

# Optional Export to excel
#rc.to_excel(cwd+"/annotated_clinvar_filt_retnet_processed.xlsx", index=False)

### MutScore Thresholds

# Likely Pathogenic thresholds
lp_scores, lp = likely_pathogenic(min_score=0, max_score=1, df=rc, classifier="MutScore")
display(lp_scores.head())
print("The upper likely-pathogenic threshold of MutScore is {}".format(lp))
#lp_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-pathogenic_MutScore.xlsx", index=False) #optional export to excel

# Likely Benign thresholds
lb_scores, lb = likely_benign(min_score=0, max_score=1, df=rc, classifier="MutScore")
display(lb_scores.head())
print("The lower likely-benign threshold of MutScore is {}".format(lb))
#lb_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-benign_MutScore.xlsx", index=False) #optional export to excel

### VEST4

# Likely Pathogenic thresholds
lp_scores, lp = likely_pathogenic(min_score=0, max_score=1, df=rc, classifier="VEST4_score")
display(lp_scores.head())
print("The upper likely-pathogenic threshold of VEST4_score is {}".format(lp))
#lp_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-pathogenic_VEST4_score.xlsx", index=False) #optional export to excel

# Likely Benign thresholds
lb_scores, lb = likely_benign(min_score=0, max_score=1, df=rc, classifier="VEST4_score")
display(lb_scores.head())
print("The lower likely-benign threshold of VEST4_score is {}".format(lb))
#lb_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-benign_VEST4_score.xlsx", index=False) #optional export to excel

### REVEL

# Likely Pathogenic thresholds
lp_scores, lp = likely_pathogenic(min_score=0, max_score=1, df=rc, classifier="REVEL_score")
display(lp_scores.head())
print("The upper likely-pathogenic threshold of REVEL_score is {}".format(lp))
#lp_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-pathogenic_REVEL_score.xlsx", index=False) #optional export to excel

# Likely Benign thresholds
lb_scores, lb = likely_benign(min_score=0, max_score=1, df=rc, classifier="REVEL_score")
display(lb_scores.head())
print("The lower likely-benign threshold of REVEL_score is {}".format(lb))
#lb_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-benign_REVEL_score.xlsx", index=False) #optional export to excel

### ClinPred

# Likely Pathogenic thresholds
lp_scores, lp = likely_pathogenic(min_score=0, max_score=1, df=rc, classifier="ClinPred_score")
display(lp_scores.head())
print("The upper likely-pathogenic threshold of ClinPred_score is {}".format(lp))
#lp_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-pathogenic_ClinPred_score.xlsx", index=False) #optional export to excel

# Likely Benign thresholds
lb_scores, lb = likely_benign(min_score=0, max_score=1, df=rc, classifier="ClinPred_score")
display(lb_scores.head())
print("The lower likely-benign threshold of ClinPred_score is {}".format(lb))
#lb_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-benign_ClinPred_score.xlsx", index=False) #optional export to excel

### MetaRNN

# Likely Pathogenic thresholds
lp_scores, lp = likely_pathogenic(min_score=0, max_score=1, df=rc, classifier="MetaRNN_score")
display(lp_scores.head())
print("The upper likely-pathogenic threshold of MetaRNN_score is {}".format(lp))
#lp_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-pathogenic_MetaRNN_score.xlsx", index=False) #optional export to excel

# Likely Benign thresholds
lb_scores, lb = likely_benign(min_score=0, max_score=1, df=rc, classifier="MetaRNN_score")
display(lb_scores.head())
print("The lower likely-benign threshold of MetaRNN_score is {}".format(lb))
#lb_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-benign_MetaRNN_score.xlsx", index=False) #optional export to excel

### BayesDel_addAF_score

# Likely Pathogenic thresholds
lp_scores, lp = likely_pathogenic(min_score=-1.293, max_score=0.758, df=rc, classifier="BayesDel_addAF_score", iterations=2052)
display(lp_scores.head())
print("The upper likely-pathogenic threshold of BayesDel_addAF_score is {}".format(lp))
#lp_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-pathogenic_BayesDel_addAF_score.xlsx", index=False) #optional export to excel

# Likely Benign thresholds
lb_scores, lb = likely_benign(min_score=-1.293, max_score=0.758, df=rc, classifier="BayesDel_addAF_score", iterations=2052)
display(lb_scores.head())
print("The lower likely-benign threshold of BayesDel_addAF_score is {}".format(lb))
#lb_scores.to_excel(cwd+"/retnet_classifier_thresholds/likely-benign_BayesDel_addAF_score.xlsx", index=False) #optional export to excel
